Seperate out calorie list
